backend_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.8
pydantic[email]==2.5.0
python-dotenv==1.0.0
alembic==1.13.0
python-multipart==0.0.6
websockets==12.0
ruff==0.1.15